chore: add zod package

+++ introducao-node-secs/src/controllers/UsersController.ts
@@ -1,6 +1,7 @@
 import { PrismaClient } from "@prisma/client";
 import { Request, Response } from "express";
 import {prisma} from "../lib/prisma"
+import { AppError } from "../errors/AppError";
 
 export class UsersController  {
     public async list(_request : Request, response : Response) {
@@ -15,6 +16,8 @@ export class UsersController  {
         const user = await prisma.user.findUnique({
             where: {id}
         })
+
+        if(!user) throw new AppError('User not found', 404);     
 
         return response.status(200).json(user);
     }
@@ -33,6 +36,13 @@ export class UsersController  {
     public async update(request: Request, response: Response){
         const {id} =  request.params;
         const {name, email} = request.body;
+
+        const userExists = await prisma.user.findUnique({
+            where: {id}
+        })
+
+        if(!userExists) throw new AppError('User not found', 404); 
+
         const user = await prisma.user.update({
             where: {id},
             data: {
@@ -46,7 +56,13 @@ export class UsersController  {
     public async delete(request : Request, response : Response) {
         const { id } = request.params;
 
-        const user = await prisma.user.delete({
+        const user = await prisma.user.findUnique({
+            where: {id}
+        })
+
+        if(!user) throw new AppError('User not found', 404); 
+
+        await prisma.user.delete({
             where: {id}
         })
